完善绘画功能

+++ quick-member/src/main/java/com/quick/member/service/DialogueTopicService.java
@@ -12,7 +12,7 @@ public interface DialogueTopicService  extends IService<DialogueTopicPO> {
      * @param type
      * @return
      */
-    List<DialogueTopicPO> getDialogueTopicsByType(Integer type);
+    List<DialogueTopicPO> getDialogueTopicsByType(Long type);
 
     void insert(DialogueTopicPO topic);
